# docs/api-spec/v0.2.0/openapi.yaml
openapi: 3.1.0
info:
  title: LakeMind API
  version: 1.0.0
  description: |
    LakeMind v0.2.0 REST API — 统一认知资产管理与受控 Job Runtime。
    所有端点（除 /health）需要 Bearer Token 认证。
    X-Tenant-Id / X-Agent-Id / X-Scopes 不再由调用方声明，由 Token 解析得到。
  contact:
    name: LakeMind Team
servers:
  - url: http://localhost:10823
    description: Local development
  - url: http://lakemind-server:10823
    description: Docker internal
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/assets:
    post:
      tags: [Assets]
      summary: 创建资产
      operationId: createAsset
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [Assets]
      summary: 资产列表
      operationId: listAssets
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/AssetType'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AssetStatus'
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseAsset'
  /api/v1/assets/{asset_id}:
    get:
      tags: [Assets]
      summary: 资产详情
      operationId: getAsset
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          description: 详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Assets]
      summary: 更新资产元数据
      operationId: updateAsset
      parameters:
        - $ref: '#/components/parameters/AssetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssetRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
    delete:
      tags: [Assets]
      summary: 删除资产（异步 Operation）
      operationId: deleteAsset
      parameters:
        - $ref: '#/components/parameters/AssetId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: 异步删除已提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOperationResponse'
  /api/v1/assets/{asset_id}/bindings:
    get:
      tags: [Assets]
      summary: Binding 列表
      operationId: getAssetBindings
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          description: Binding 列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Binding'
  /api/v1/assets/{asset_id}/lineage:
    get:
      tags: [Assets]
      summary: 资产血缘
      operationId: getAssetLineage
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          description: 血缘图
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageGraph'
  /api/v1/assets/{asset_id}/reindex:
    post:
      tags: [Assets]
      summary: 重建索引（异步 Operation）
      operationId: reindexAsset
      parameters:
        - $ref: '#/components/parameters/AssetId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: 重建已提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOperationResponse'

  # ────────────── Knowledge ──────────────
...
  /api/v1/jobs:
    post:
      tags: [Jobs]
      summary: 提交 Job
      operationId: submitJob
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitJobRequest'
      responses:
        '201':
          description: 提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRun'
    get:
      tags: [Jobs]
      summary: Job 列表
      operationId: listJobs
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/JobStatus'
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseJobRun'
  /api/v1/jobs/{job_id}:
    get:
      tags: [Jobs]
      summary: Job 详情
      operationId: getJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: 详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRun'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/jobs/{job_id}/cancel:
    post:
      tags: [Jobs]
      summary: 取消 Job
      operationId: cancelJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: 取消成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRun'
  /api/v1/jobs/{job_id}/retry:
    post:
      tags: [Jobs]
      summary: 重试 Job（新 Attempt）
      operationId: retryJob
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 重试已提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRun'
  /api/v1/jobs/{job_id}/result:
    get:
      tags: [Jobs]
      summary: Job 结果
      operationId: getJobResult
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: 结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artifact'
  /api/v1/jobs/{job_id}/attempts:
    get:
      tags: [Jobs]
      summary: Job Attempt 列表
      operationId: getJobAttempts
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Attempt 列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobAttempt'

  # ────────────── Operations ──────────────
...
  /api/v1/configuration/{scope}:
    get:
      tags: [Configuration]
      summary: 按作用域获取配置
      operationId: getConfig
      parameters:
        - name: scope
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 配置值
          content:
            application/json:
              schema:
                type: object
    put:
      tags: [Configuration]
      summary: 设置配置（产生 Revision）
      operationId: setConfig
      parameters:
        - name: scope
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
                value: {}
                reason:
                  type: string
              required: [key, value, reason]
      responses:
        '201':
          description: Revision 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigRevision'
...
  /api/v1/configuration/revisions/{rev_id}/activate:
    post:
      tags: [Configuration]
      summary: 激活 Revision
      operationId: activateConfigRevision
      parameters:
        - $ref: '#/components/parameters/RevisionId'
      responses:
        '202':
          description: 激活已提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOperationResponse'
  /api/v1/configuration/revisions/{rev_id}/rollback:
    post:
      tags: [Configuration]
      summary: 回滚 Revision
      operationId: rollbackConfigRevision
      parameters:
        - $ref: '#/components/parameters/RevisionId'
      responses:
        '202':
          description: 回滚已提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOperationResponse'

  # ────────────── Security ──────────────
  /api/v1/security/tokens:
    post:
      tags: [Security]
      summary: 签发 Token
      operationId: issueToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueTokenRequest'
      responses:
        '201':
          description: 签发成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
    get:
      tags: [Security]
      summary: Token 列表
      operationId: listTokens
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Token'
  /api/v1/security/tokens/{token_id}:
    delete:
      tags: [Security]
      summary: 撤销 Token
      operationId: revokeToken
      parameters:
        - name: token_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 撤销成功
...
  /api/v1/audit:
    get:
      tags: [Audit]
      summary: 查询审计事件
      operationId: queryAudit
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: event_type
          in: query
          schema:
            type: string
        - name: principal
          in: query
          schema:
            type: string
        - name: resource
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 审计事件列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseAuditEvent'
  /api/v1/audit/export:
    get:
      tags: [Audit]
      summary: 导出审计事件
      operationId: exportAudit
      parameters:
        - name: event_type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 导出流
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  # ────────────── Instances ──────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: 幂等键（24h 内相同 Principal + 相同 Key 返回缓存响应）
    AssetId:
      name: asset_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^ast_[0-9A-HJKMNP-TV-Z]{26}$'
...
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^job_[0-9A-HJKMNP-TV-Z]{26}$'
...
    RevisionId:
      name: rev_id
      in: path
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: 校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - AUTHENTICATION_FAILED
                - TOKEN_REVOKED
                - PERMISSION_DENIED
                - TENANT_SCOPE_VIOLATION
                - RESOURCE_NOT_FOUND
                - ASSET_NOT_READY
                - ASSET_DEGRADED
                - ASSET_FAILED
                - SKILL_NOT_PUBLISHED
                - SKILL_VERSION_IMMUTABLE
                - JOB_RESOURCE_DENIED
                - JOB_NOT_FOUND
                - MODEL_DEPLOYMENT_UNAVAILABLE
                - EMBEDDING_SPACE_MISMATCH
                - CONFIG_REVISION_CONFLICT
                - OPERATION_APPROVAL_REQUIRED
                - OPERATION_NOT_APPROVED
                - IDEMPOTENCY_CONFLICT
                - VALIDATION_FAILED
                - RATE_LIMITED
                - INTERNAL_ERROR
                - UPSTREAM_TIMEOUT
            message:
              type: string
            request_id:
              type: string
            resource_id:
              type: string
            resource_status:
              type: string
            retryable:
              type: boolean
            details:
              type: object
            correlation_id:
              type: string
    AssetType:
      type: string
      enum: [knowledge, skill, memory]
    AssetStatus:
      type: string
      enum: [CREATING, PROCESSING, READY, DEGRADED, FAILED, DELETED]
    JobStatus:
      type: string
      enum: [PENDING, RUNNING, SUCCEEDED, FAILED, CANCELLED, LOST]
...
    Asset:
      type: object
      properties:
        asset_id:
          type: string
        tenant_id:
          type: string
        type:
          $ref: '#/components/schemas/AssetType'
        name:
          type: string
        status:
          $ref: '#/components/schemas/AssetStatus'
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAssetRequest:
      type: object
      required: [type, name, source]
      properties:
        type:
          $ref: '#/components/schemas/AssetType'
        name:
          type: string
        source:
          type: object
        metadata:
          type: object
    UpdateAssetRequest:
      type: object
      properties:
        metadata:
          type: object
    Binding:
      type: object
      properties:
        binding_id:
          type: string
        asset_id:
          type: string
        kind:
          type: string
        status:
          type: string
        required:
          type: boolean
    LineageGraph:
      type: object
      properties:
        nodes:
          type: array
          items:
            type: object
        edges:
          type: array
          items:
            type: object
...
    JobRun:
      type: object
      properties:
        job_id:
          type: string
        tenant_id:
          type: string
        skill_ref:
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
        params:
          type: object
        created_at:
          type: string
          format: date-time
    SubmitJobRequest:
      type: object
      required: [skill_ref, inputs]
      properties:
        skill_ref:
          type: string
        inputs:
          type: object
        params:
          type: object
        model_profile:
          type: string
        resource_overrides:
          type: object
    JobAttempt:
      type: object
      properties:
        attempt_id:
          type: string
        job_id:
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
    Artifact:
      type: object
      properties:
        artifact_id:
          type: string
        job_id:
          type: string
        uri:
          type: string
        size:
          type: integer
...
    AsyncOperationResponse:
      type: object
      properties:
        operation_id:
          type: string
        resource_id:
          type: string
        status:
          type: string
        message:
          type: string
...
    ConfigRevision:
      type: object
      properties:
        revision_id:
          type: string
        scope:
          type: string
        status:
          type: string
          enum: [DRAFT, ACTIVE, ROLLED_BACK]
        created_at:
          type: string
          format: date-time
    Token:
      type: object
      properties:
        token_id:
          type: string
        token:
          type: string
        agent_id:
          type: string
        tenant_id:
          type: string
        scopes:
          type: array
          items:
            type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
    IssueTokenRequest:
      type: object
      required: [agent_id, tenant_id, scopes]
      properties:
        agent_id:
          type: string
        tenant_id:
          type: string
        scopes:
          type: array
          items:
            type: string
...
    AuditEvent:
      type: object
      properties:
        event_id:
          type: string
        event_type:
          type: string
        principal:
          type: string
        resource:
          type: string
        action:
          type: string
        result:
          type: string
        details:
          type: object
        created_at:
          type: string
          format: date-time
...
    PageResponseAsset:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Asset'
...
    PageResponseJobRun:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/JobRun'
...
    PageResponseAuditEvent:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AuditEvent'
    PageResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        has_next:
          type: boolean